Revert "refactoring to follow the ports and adapters pattern"
