--- src/main/java/com/reimbursement/approval_service/adapters/inbound/views/ApprovalView.java
@@ -1,9 +1,0 @@
-package com.reimbursement.approval_service.adapters.inbound.views;
-
-public class ApprovalView {
-    public interface CreateApproval {
-    }
-
-    public interface UpdateApproval {
-    }
-}
