--- src/main/java/com/reimbursement/approval_service/application/services/ExpenseTopicService.java
@@ -1,13 +1,0 @@
-package com.reimbursement.approval_service.application.services;
-
-import org.eclipse.paho.client.mqttv3.MqttException;
-
-import com.reimbursement.approval_service.adapters.inbound.dtos.ApprovalDto;
-
-public interface ExpenseTopicService {
-
-    void subscribe() throws MqttException;
-
-    void publishMessage(ApprovalDto approval) throws MqttException;
-
-}
