--- src/main/java/com/reimbursement/approval_service/application/specifications/ApprovalSpecification.java
@@ -1,17 +1,0 @@
-package com.reimbursement.approval_service.application.specifications;
-
-import org.springframework.data.jpa.domain.Specification;
-
-import com.reimbursement.approval_service.domain.Approval.Approval;
-
-import net.kaczmarzyk.spring.data.jpa.domain.Equal;
-import net.kaczmarzyk.spring.data.jpa.domain.NotNull;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
-
-@And({
-                @Spec(path = "deleted_at", params = "isDeleted", spec = NotNull.class),
-                @Spec(path = "status", spec = Equal.class)
-})
-public interface ApprovalSpecification extends Specification<Approval> {
-}
