+++ src/main/java/com/reimbursement/approval_service/clients/ExpensesRequest.java
@@ -1,0 +1,16 @@
+package com.reimbursement.approval_service.clients;
+
+import com.reimbursement.approval_service.enums.Status;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ExpensesRequest {
+
+    private Status status;
+
+}
