+++ src/main/java/com/reimbursement/approval_service/config/AppSetup.java
@@ -1,0 +1,21 @@
+package com.reimbursement.approval_service.config;
+
+import java.util.List;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import io.micrometer.common.lang.NonNull;
+import net.kaczmarzyk.spring.data.jpa.web.SpecificationArgumentResolver;
+
+@Configuration
+public class AppSetup implements WebMvcConfigurer {
+
+    @Override
+    public void addArgumentResolvers(@NonNull List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new SpecificationArgumentResolver());
+        resolvers.add(new PageableHandlerMethodArgumentResolver());
+    }
+}
