--- src/main/java/com/reimbursement/approval_service/domain/Approval/ApprovalStatus.java
@@ -1,7 +1,0 @@
-package com.reimbursement.approval_service.domain.Approval;
-
-public enum ApprovalStatus {
-    PENDING,
-    APPROVED,
-    REJECTED
-}
