+++ src/main/java/com/reimbursement/approval_service/enums/Status.java
@@ -1,0 +1,7 @@
+package com.reimbursement.approval_service.enums;
+
+public enum Status {
+    PENDING,
+    APPROVED,
+    REJECTED
+}
