+++ src/main/java/com/reimbursement/approval_service/exceptions/ResourceDuplicated.java
@@ -1,0 +1,8 @@
+package com.reimbursement.approval_service.exceptions;
+
+public class ResourceDuplicated extends RuntimeException {
+    public ResourceDuplicated(String message) {
+        super(message);
+    }
+
+}
