+++ src/main/java/com/reimbursement/approval_service/exceptions/ResourceNotFound.java
@@ -1,0 +1,7 @@
+package com.reimbursement.approval_service.exceptions;
+
+public class ResourceNotFound extends RuntimeException {
+    public ResourceNotFound(String message) {
+        super(message);
+    }
+}
