--- src/main/java/com/reimbursement/approval_service/infrastructure/config/AppConfig.java
@@ -1,21 +1,0 @@
-package com.reimbursement.approval_service.infrastructure.config;
-
-import java.util.List;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.web.PageableHandlerMethodArgumentResolver;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import net.kaczmarzyk.spring.data.jpa.web.SpecificationArgumentResolver;
-
-@Configuration
-public class AppConfig implements WebMvcConfigurer {
-
-    @SuppressWarnings("null")
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new SpecificationArgumentResolver());
-        resolvers.add(new PageableHandlerMethodArgumentResolver());
-    }
-}
