--- src/main/java/com/reimbursement/approval_service/infrastructure/config/MqttConfig.java
@@ -1,29 +1,0 @@
-package com.reimbursement.approval_service.infrastructure.config;
-
-import org.eclipse.paho.client.mqttv3.MqttClient;
-import org.eclipse.paho.client.mqttv3.MqttConnectOptions;
-import org.eclipse.paho.client.mqttv3.MqttException;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class MqttConfig {
-
-    private final String broker = "tcp://localhost:1883";
-    private final String clientId = "approval-service-client";
-
-    @Bean
-    public MqttClient mqttClient() throws MqttException {
-        MqttClient client = new MqttClient(broker, clientId);
-
-        MqttConnectOptions options = new MqttConnectOptions();
-        options.setAutomaticReconnect(true);
-        options.setCleanSession(true);
-        options.setConnectionTimeout(10);
-
-        client.connect(options);
-
-        return client;
-    }
-
-}
