--- src/main/java/com/reimbursement/approval_service/infrastructure/exceptions/ResourceDuplicated.java
@@ -1,8 +1,0 @@
-package com.reimbursement.approval_service.infrastructure.exceptions;
-
-public class ResourceDuplicated extends RuntimeException {
-    public ResourceDuplicated(String message) {
-        super(message);
-    }
-
-}
