--- src/main/java/com/reimbursement/approval_service/infrastructure/exceptions/ResourceNotFound.java
@@ -1,7 +1,0 @@
-package com.reimbursement.approval_service.infrastructure.exceptions;
-
-public class ResourceNotFound extends RuntimeException {
-    public ResourceNotFound(String message) {
-        super(message);
-    }
-}
