+++ src/main/java/com/reimbursement/approval_service/security/ApplicationSecurity.java
@@ -1,0 +1,39 @@
+package com.reimbursement.approval_service.security;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.Customizer;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class ApplicationSecurity {
+
+    @Value("${spring.security.oauth2.resourceserver.jwt.jwk-set-uri}")
+    String jwkSetUri;
+
+    @SuppressWarnings("unused")
+    SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf(csrf -> csrf.disable())
+                .authorizeHttpRequests(req -> {
+                    req.anyRequest().authenticated();
+                })
+                .oauth2ResourceServer(oauth2 -> {
+                    oauth2.jwt(Customizer.withDefaults());
+                });
+
+        return http.build();
+    }
+
+    @Bean
+    @SuppressWarnings("unused")
+    JwtDecoder jwtDecoder() {
+        return NimbusJwtDecoder.withJwkSetUri(this.jwkSetUri).build();
+    }
+}
