+++ src/main/java/com/reimbursement/approval_service/views/ApprovalView.java
@@ -1,0 +1,9 @@
+package com.reimbursement.approval_service.views;
+
+public class ApprovalView {
+    public interface CreateApproval {
+    }
+
+    public interface UpdateApproval {
+    }
+}
